EquipmentSkinnedMesh.cs
It just works now
You dont need to configure anything for this to work now
