--- Assets/KoboldKare/Scripts/SkinnedMeshEquipmentShapeCopier.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: a5614f23e236e5d41b7bb6d7ccc2a384
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
